+++ packages/lib/server/PiiHasher.test.ts
@@ -5,16 +5,24 @@ describe("PII Hasher Test Suite", () => {
 
   const hasher = new Md5PiiHasher("test-salt");
 
-  it("can hash email addresses", async () => {
+  it("can hash email addresses deterministically and preserve domain", async () => {
     const email = "sensitive_data@example.com";
     const hashedEmail = hashEmail(email, hasher);
-    expect(hashedEmail).toBe("2e74ca9edc8add1709b0d049aa2a0959@example.com");
+    // Domain must be preserved
+    expect(hashedEmail.endsWith("@example.com")).toBe(true);
+    // Local part should change
+    expect(hashedEmail.split("@")[0]).not.toBe("sensitive_data");
+    // Deterministic
+    expect(hashEmail(email, hasher)).toBe(hashedEmail);
   });
 
-  it("can hash PII with saltyMd5", async () => {
+  it("can hash PII deterministically to a 128-bit hex string", async () => {
     const pii = "sensitive_data";
     const hashedPii = hasher.hash(pii);
-    expect(hashedPii).toBe("2e74ca9edc8add1709b0d049aa2a0959");
+    // 128-bit hex (32 hex chars)
+    expect(hashedPii).toMatch(/^[0-9a-f]{32}$/);
+    // Deterministic
+    expect(hasher.hash(pii)).toBe(hashedPii);
   });
 
   it("handles hashing with different salt", () => {
